requirements.txt
# Core dependencies
torch>=1.9.0
numpy>=1.19.0
gymnasium>=0.26.0
dm_control>=1.0.0

# Configuration and logging
pyyaml>=5.4.0
loguru>=0.5.0

# Optional but recommended
wandb>=0.12.0

# DMC datasets from GitHub
git+https://github.com/davidireland3/dmc_datasets.git@main
